Implementerede SeparationCheckerTest og VelocityCheckerTest. Tilføjede tests til TrackObjectTest.

+++ SWT3/PrintDataFromDLL/ATMClasses/Interfaces/IVelocityCourseCalculator.cs
@@ -14,7 +14,7 @@ namespace ATMClasses.Interfaces
         
 
         //Returns velocity in whole meters per second
-        int CalculateVelocity(TrackObject oldTO, TrackObject newTO);
+        Int64 CalculateVelocity(TrackObject oldTO, TrackObject newTO);
 
     }
 }
